+++ R/residuals.mvgam.R
@@ -51,6 +51,13 @@ residuals.mvgam <- function(
   probs = c(0.025, 0.975),
   ...
 ) {
+
+  if (length(probs) != 2L) {
+    stop("argument 'probs' must be a vector of length 2", call. = FALSE)
+  }
+  validate_proportional(min(probs))
+  validate_proportional(max(probs))
+
   # What was the original time / series order?
   orig_order <- data.frame(
     series = object$obs_data$series,